#15 Create InvoiceBook class. Validated arguments in InvoiceBook.

+++ src/main/java/pl/coderstrust/logic/InvoiceBook.java
@@ -20,6 +20,12 @@ public class InvoiceBook {
   }
 
   public List<Invoice> getAllInvoicesInGivenDateRange(LocalDate fromDate, LocalDate toDate) throws InvoiceRepositoryOperationException {
+    if (fromDate == null) {
+      throw new IllegalArgumentException("FromDate cannot be null");
+    }
+    if (toDate == null) {
+      throw new IllegalArgumentException("ToDate cannot be null");
+    }
     List<Invoice> invoices = getAllInvoices();
     List<Invoice> invoicesInGivenDateRange = new ArrayList<>(invoices);
     for (Invoice invoice :
@@ -32,22 +38,37 @@ public class InvoiceBook {
   }
 
   public Invoice getInvoiceById(Integer id) throws InvoiceRepositoryOperationException {
+    if (id < 0) {
+      throw new IllegalArgumentException("Id cannot be less then zero.");
+    }
     return invoiceRepository.findById(id);
   }
 
   public Invoice addInvoice(Invoice invoice) throws InvoiceRepositoryOperationException {
+    if (invoice == null) {
+      throw new IllegalArgumentException("Invoice cannot be null.");
+    }
     return invoiceRepository.save(invoice);
   }
 
   public void updateInvoice(Invoice invoice) throws InvoiceRepositoryOperationException {
+    if (invoice == null) {
+      throw new IllegalArgumentException("Invoice cannot be null.");
+    }
     invoiceRepository.save(invoice);
   }
 
   public void deleteInvoiceById(Integer id) throws InvoiceRepositoryOperationException {
+    if (id < 0) {
+      throw new IllegalArgumentException("Id cannot be less then zero.");
+    }
     invoiceRepository.deleteById(id);
   }
 
   public void deleteInvoice(Invoice invoice) throws InvoiceRepositoryOperationException {
+    if (invoice == null) {
+      throw new IllegalArgumentException("Invoice cannot be null.");
+    }
     if (invoiceRepository.existsById(invoice.getId())) {
       deleteInvoiceById(invoice.getId());
     }
